+++ application/services/Etablissement.php
@@ -1342,4 +1342,9 @@ class Service_Etablissement implements Service_Interface_Etablissement
             return "avisDiff";
         }
     }
+    
+    public function getDossierDonnantAvis($idEtablissement) {
+        $DBEtab = new Model_DbTable_Etablissement;
+        return $DBEtab->getDossierDonnantAvis($idEtablissement);
+    }
 }